extract functions from tedgeBackend -> tedgeFileStore
